# fms-ui/src/assets/fms-api-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Fleet Management System API Documentation",
    "description": "Dokumentacja API - autoryzacja za pomocą JWT",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/api/v1/manager/updateVehicle/{vehicleId}": {
      "put": {
        "tags": [
          "manager-controller"
        ],
        "operationId": "updateVehicle",
        "parameters": [
          {
            "name": "vehicleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVehicleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/manager/newService/{vehicleId}": {
      "post": {
        "tags": [
          "manager-controller"
        ],
        "operationId": "newService",
        "parameters": [
          {
            "name": "vehicleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewServiceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/manager/newMileage/{vehicleId}": {
      "post": {
        "tags": [
          "manager-controller"
        ],
        "operationId": "newMileage",
        "parameters": [
          {
            "name": "vehicleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewMileageChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MileageChangeDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/manager/newInsurance/{vehicleId}": {
      "post": {
        "tags": [
          "manager-controller"
        ],
        "operationId": "newInsurance",
        "parameters": [
          {
            "name": "vehicleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewInsuranceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/InsuranceDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/manager/newInspection/{vehicleId}": {
      "post": {
        "tags": [
          "manager-controller"
        ],
        "operationId": "newInspection",
        "parameters": [
          {
            "name": "vehicleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewInspectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/manager/newAssignment/{userId}/{vehicleId}": {
      "post": {
        "tags": [
          "manager-controller"
        ],
        "operationId": "newAssignment",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "vehicleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/manager/createVehicle": {
      "post": {
        "tags": [
          "manager-controller"
        ],
        "operationId": "createVehicle",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVehicleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CreateVehicleResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/manager/updateService/{serviceId}": {
      "patch": {
        "tags": [
          "manager-controller"
        ],
        "operationId": "updateService",
        "parameters": [
          {
            "name": "serviceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServiceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/manager/updateInsurance/{insuranceId}": {
      "patch": {
        "tags": [
          "manager-controller"
        ],
        "operationId": "updateInsurance",
        "parameters": [
          {
            "name": "insuranceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInsuranceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/InsuranceDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/manager/updateInspection/{inspectionId}": {
      "patch": {
        "tags": [
          "manager-controller"
        ],
        "operationId": "updateInspection",
        "parameters": [
          {
            "name": "inspectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInspectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/manager/updateCost/{costId}": {
      "patch": {
        "tags": [
          "manager-controller"
        ],
        "operationId": "updateCost",
        "parameters": [
          {
            "name": "costId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CostDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/manager/endAssignment/{assignmentId}": {
      "patch": {
        "tags": [
          "manager-controller"
        ],
        "operationId": "endAssignment",
        "parameters": [
          {
            "name": "assignmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/manager/changeStatusForVehicle/{vehicleId}": {
      "patch": {
        "tags": [
          "manager-controller"
        ],
        "operationId": "changeStatusForVehicle",
        "parameters": [
          {
            "name": "vehicleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeStatusResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/manager/getUserAssignments/{userId}": {
      "get": {
        "tags": [
          "manager-controller"
        ],
        "operationId": "getUserAssignments",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserAssignmentResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/manager/getServicesForVehicle/{vehicleId}": {
      "get": {
        "tags": [
          "manager-controller"
        ],
        "operationId": "getServicesForVehicle",
        "parameters": [
          {
            "name": "vehicleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServiceDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/manager/getServiceById/{serviceId}": {
      "get": {
        "tags": [
          "manager-controller"
        ],
        "operationId": "getServiceById",
        "parameters": [
          {
            "name": "serviceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/manager/getRefuelingsForVehicle/{vehicleId}": {
      "get": {
        "tags": [
          "manager-controller"
        ],
        "operationId": "getRefuelingsForVehicle",
        "parameters": [
          {
            "name": "vehicleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RefuelingDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/manager/getRefuelingById/{refuelingId}": {
      "get": {
        "tags": [
          "manager-controller"
        ],
        "operationId": "getRefuelingById",
        "parameters": [
          {
            "name": "refuelingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RefuelingDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/manager/getIotConfig/{vehicleId}": {
      "get": {
        "tags": [
          "manager-controller"
        ],
        "operationId": "getIotConfig",
        "parameters": [
          {
            "name": "vehicleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "byte"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/manager/getInsuranceById/{insuranceId}": {
      "get": {
        "tags": [
          "manager-controller"
        ],
        "operationId": "getInsuranceById",
        "parameters": [
          {
            "name": "insuranceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/InsuranceDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/manager/getInspectionsForVehicle/{vehicleId}": {
      "get": {
        "tags": [
          "manager-controller"
        ],
        "operationId": "getInspectionsForVehicle",
        "parameters": [
          {
            "name": "vehicleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InspectionDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/manager/getInspectionById/{inspectionId}": {
      "get": {
        "tags": [
          "manager-controller"
        ],
        "operationId": "getInspectionById",
        "parameters": [
          {
            "name": "inspectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/manager/getCostById/{costId}": {
      "get": {
        "tags": [
          "manager-controller"
        ],
        "operationId": "getCostById",
        "parameters": [
          {
            "name": "costId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CostDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/manager/getAssignmentsForVehicle/{vehicleId}": {
      "get": {
        "tags": [
          "manager-controller"
        ],
        "operationId": "getAssignmentsForVehicle",
        "parameters": [
          {
            "name": "vehicleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssignmentDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/manager/getAssignmentById/{assignmentId}": {
      "get": {
        "tags": [
          "manager-controller"
        ],
        "operationId": "getAssignmentById",
        "parameters": [
          {
            "name": "assignmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/manager/getAllInsurancesForVehicle/{vehicleId}": {
      "get": {
        "tags": [
          "manager-controller"
        ],
        "operationId": "getAllInsurancesForVehicle",
        "parameters": [
          {
            "name": "vehicleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InsuranceDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/manager/getAllCostsForVehicle": {
      "get": {
        "tags": [
          "manager-controller"
        ],
        "operationId": "getAllCostsForVehicle",
        "parameters": [
          {
            "name": "vehicleId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CostDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/manager/deleteVehicle/{vehicleId}": {
      "delete": {
        "tags": [
          "manager-controller"
        ],
        "operationId": "deleteVehicle",
        "parameters": [
          {
            "name": "vehicleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UpdateVehicleRequest": {
        "type": "object",
        "properties": {
          "manufacturer": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "registrationNumber": {
            "type": "string"
          },
          "vin": {
            "type": "string"
          },
          "fuelType": {
            "type": "string",
            "enum": [
              "GASOLINE",
              "DIESEL",
              "ELECTRIC",
              "DIESEL_HYBRID",
              "GASOLINE_HYBRID",
              "LPG"
            ]
          },
          "displacement": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "MileageChangeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "newMileage": {
            "type": "integer",
            "format": "int32"
          },
          "vehicleId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ServiceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "mileageAtTheTime": {
            "type": "integer",
            "format": "int32"
          },
          "vehicleId": {
            "type": "integer",
            "format": "int32"
          },
          "cost": {
            "$ref": "#/components/schemas/CostDto"
          }
        }
      },
...
      "NewServiceRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "mileageAtTheTime": {
            "type": "integer",
            "format": "int32"
          },
          "cost": {
            "type": "number"
          }
        }
      },
      "NewMileageChangeRequest": {
        "type": "object",
        "properties": {
          "newMileage": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}